+++ packages/atomsbox/lib/src/atomsbox/atoms/app_outlined_button.dart
@@ -151,7 +151,7 @@ final appOutlinedButtonThemeDark = OutlinedButtonThemeData(
       ),
     ),
     side: BorderSide(color: AppColors.appColorSchemeDark.primary),
-    minimumSize: const Size(100, 48),
+    minimumSize: const Size(100, 36),
     foregroundColor: AppColors.appColorSchemeDark.primary,
     backgroundColor: Colors.transparent,
     disabledForegroundColor:
@@ -172,7 +172,7 @@ final appOutlinedButtonThemeLight = OutlinedButtonThemeData(
       color: AppColors.appColorSchemeLight.primary,
       strokeAlign: BorderSide.strokeAlignCenter,
     ),
-    minimumSize: const Size(100, 48),
+    minimumSize: const Size(100, 36),
     foregroundColor: AppColors.appColorSchemeLight.primary,
     backgroundColor: Colors.transparent,
     disabledForegroundColor: