+++ packages/atomsbox/lib/src/atomsbox/organisms/organisms.dart
@@ -7,3 +7,4 @@ export 'app_form.dart';
 export 'app_grid.dart';
 export 'app_list.dart';
 export 'app_tab.dart';
+export 'app_tree_view.dart';