+++ packages/atomsbox/CHANGELOG.md
@@ -1,2 +1,9 @@
+## 0.2.0
+* New components: AppRatingBar, AppTreeView
+* Add surfaceTintColor to customize the AppCard
+* Replace atomsbox logo in the README file
+* Remove just_audio and audio_service, add flutter_rating_bar
+* Set buttons default height to 36
+
 ## 0.1.0
 * Initial version of the atomsbox component library.